+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ca-pi",
-  "version": "0.1.42",
+  "version": "0.1.43",
   "private": true,
   "license": "AGPL-3.0-only",
   "engines": {
